requirements.txt
langchain
langchain_community
numpy 
pandas
faiss-cpu
dotenv
euriai
streamlit